Added the ability to move between rooms
Added portal objects which allow the player to advance to the next room. Added a portal to each room. Changed the handling of player movement to allow for the proper functioning of portals.

Added better handling of FileNotFoundException by having it handled within the file-based Room constructor.

Fixed the locating of the spawn point in the file-based Room constructor, as the coordinates were swapped.

Added the ability to win the game once the player has gone through all fo the rooms.

Removed the uneeded, commented out code in the Main class.